incorporate BaseLogger
